Came back to previous commit to solve some issues. Now they're solved and everything works fine

+++ app/Http/Livewire/PhaseController.php
@@ -5,8 +5,10 @@ namespace App\Http\Livewire;
 use App\Models\Phase;
 use App\Models\Project;
 use App\Models\Task;
+use App\Models\User;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Validation\Rule;
 use Livewire\Component;
 
 class PhaseController extends Component
@@ -20,8 +22,8 @@ class PhaseController extends Component
     public $editPhase = false;
 
     public $title;
-    public $start_time;
-    public $end_time;
+    public $start_date;
+    public $end_date;
     public $hour_estimate;
     public $content;
     public $priority;
@@ -42,9 +44,17 @@ class PhaseController extends Component
     {
         $rules = [
             "title" => ['required', 'string', 'max:255'],
-            "start_time" => ['required', 'date', 'after_or_equal:today'],
-            "end_time" => ['required', 'date', 'after_or_equal:start_time'],
-            "hour_estimate" => ['required', 'integer', 'between:0,100.99'],
+            "start_date" => [
+                Rule::when($this->phase && $this->start_date != $this->phase->start_date, function () {
+                    return ['required', 'date', 'after_or_equal:today'];
+                }),
+            ],
+            "end_date" => [
+                Rule::when($this->phase, function () {
+                    return ['required', 'date', 'after_or_equal:start_date'];
+                }),
+            ],
+            "hour_estimate" => ['required', 'integer', 'min:1'],
             "content" => ['required', 'string', 'max:500'],
             "priority" => ['required', 'in:Low,Medium,High,Urgent'],
             'project_id' => 'required',
@@ -63,35 +73,45 @@ class PhaseController extends Component
     public function render()
     {
         $projects = Project::where('leader_id_assigned', Auth::user()->id)->get();
-
+        
         $phases = Phase::with('project')
                     ->where('user_id', Auth::user()->id)
                     ->where('title', 'like', '%'.$this->search.'%')
                     ->orderBy('order_position', 'asc')
                     ->get();
+        
+        $adminUsers = User::role('admin-user')->get();
+        if($adminUsers->contains(auth()->user())){
+            $projects = Project::all();
+            $phases = Phase::with('project')
+                    ->where('title', 'like', '%'.$this->search.'%')
+                    ->orderBy('order_position', 'asc')
+                    ->get();
+        }
 
         $groupedPhases = $phases->groupBy('project_id');
 
-        return view('livewire.phase', ['projects' => $projects, 'groupedPhases' => $groupedPhases, ])->layout('layouts.app');
+        return view('livewire.phase', ['projects' => $projects, 'groupedPhases' => $groupedPhases ])->layout('layouts.app');
     }
 
     public function setValues($id)
     {
         $this->phase = Phase::find($id);
         $this->title = $this->phase->title;
-        $this->start_time = $this->phase->start_time;
-        $this->end_time = $this->phase->end_time;
+        $this->start_date = $this->phase->start_date;
+        $this->end_date = $this->phase->end_date;
         $this->hour_estimate = $this->phase->hour_estimate;
         $this->content = $this->phase->content;
         $this->priority = $this->phase->priority;
+        $this->project_id = $this->phase->project_id;
     }
 
     public function resetValues()
     {
         $this->phase = new Phase();
         $this->title = "";
-        $this->start_time = now()->format('Y-m-d');
-        $this->end_time = "";
+        $this->start_date = now()->format('Y-m-d');
+        $this->end_date = "";
         $this->hour_estimate = "";
         $this->content = "";
         $this->priority = null;
@@ -105,13 +125,6 @@ class PhaseController extends Component
         $this->openModal = true;
     }
 
-    public function editPhaseNote($id)
-    {
-        $this->setValues($id);
-        $this->editPhase = true;
-        $this->openModal = true;
-    }
-
     public function savePhase()
     {
         $this->validate();
@@ -119,15 +132,21 @@ class PhaseController extends Component
         $this->phase = new Phase();
         $this->phase->user_id = Auth::user()->id;
         $this->phase->title = $this->title;
-        $this->phase->start_time = $this->start_time;
-        $this->phase->end_time = $this->end_time;
+        $this->phase->start_date = $this->start_date;
+        $this->phase->end_date = $this->end_date;
         $this->phase->hour_estimate = $this->hour_estimate;
         $this->phase->content = $this->content;
         $this->phase->priority = $this->priority;
         $this->phase->project_id = $this->project_id;
         $this->phase->save();
         $this->openModal = false;
-        return redirect()->route('phases');
+    }
+
+    public function editPhaseNote($id)
+    {
+        $this->setValues($id);
+        $this->editPhase = true;
+        $this->openModal = true;
     }
 
     public function editPhase($id)
@@ -137,8 +156,8 @@ class PhaseController extends Component
         $this->phase = Phase::find($id);
         $this->phase->user_id = Auth::user()->id;
         $this->phase->title = $this->title;
-        $this->phase->start_time = $this->start_time;
-        $this->phase->end_time = $this->end_time;
+        $this->phase->start_date = $this->start_date;
+        $this->phase->end_date = $this->end_date;
         $this->phase->hour_estimate = $this->hour_estimate;
         $this->phase->content = $this->content;
         $this->phase->priority = $this->priority;
@@ -148,9 +167,14 @@ class PhaseController extends Component
 
     public function deletePhase($id)
     {
-        Phase::destroy($id);
-        $this->openModal = false;
-        return redirect()->route('phases');
+        if(Task::where('phase_id', $id)->count() == 0){
+            Phase::destroy($id);
+            $this->openModal = false;
+            return redirect()->route('phases');
+        }
+        else{
+            $this->emit('alert', "You can't delete a phase that already has tasks assigned", route('phases'));
+        }
     }
 
     public function updateTaskOrder($items)
@@ -165,12 +189,20 @@ class PhaseController extends Component
 
     public function getProgressPercentage($id)
     {
+        $phase = Phase::find($id);
         if(Task::where('phase_id', $id)->count() > 0)
         {
             $tasks = Task::where('phase_id', $id)->get();
             $completedTasksCount = $tasks->filter(function ($task) {
                 return $task->is_finished;
             })->count();
+            if($completedTasksCount == $phase->task()->count()){
+                $phase->is_finished = true;
+            }
+            else{
+                $phase->is_finished = false;
+            }
+            $phase->save();
             $percentage = ($completedTasksCount / $tasks->count()) * 100;
             return round($percentage) . '%';
         }
